@@ -1,6 +1,5 @@
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:custom_pop_up_menu/custom_pop_up_menu.dart';
-import 'package:dostop_v2/src/providers/codigos_residente_provider.dart';
 import 'package:dostop_v2/src/providers/config_usuario_provider.dart';
 import 'package:dostop_v2/src/providers/login_provider.dart';
 import 'package:dostop_v2/src/widgets/custom_qr.dart';
@@ -22,9 +21,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
 
 import '../../main.dart';
+import '../providers/codigos_residente_provider.dart';
+import '../widgets/custom_qr.dart';
 
 class HomePage extends StatefulWidget {
-  final PageController pageController;
+  final PageController? pageController;
   HomePage({this.pageController});
   @override
   _HomePageState createState() => _HomePageState();
@@ -41,27 +42,26 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
 
   CustomPopupMenuController _controller = CustomPopupMenuController();
   bool _nuevaEncuesta = false, _accesos = false, _qrResidente = false;
-  EncuestaModel _datosEncuesta;
+  EncuestaModel? _datosEncuesta;
   int _noMolestar = 2;
   String _numeroCaseta = '';
 
   void _obtenerEncuesta() {
     avisosProvider.obtenerUltimaEncuesta(_prefs.usuarioLogged).then((encuesta) {
-      ///previene la llamada del setState cuando el widget ya ha sido destruido. (if (!mounted) return;)
       if (!mounted) return;
       setState(() {
         if (encuesta.containsKey(1) && !_nuevaEncuesta) {
           _nuevaEncuesta = true;
           _datosEncuesta = encuesta[1];
-          _scaffoldKey.currentState
+          ScaffoldMessenger.of(context)
               .showSnackBar(SnackBar(
                 duration: Duration(minutes: 1),
                 content: Text(
                     'Hay disponible una nueva encuesta\n¡Nos interesa tu opinión!'),
                 action: SnackBarAction(
                   onPressed: () => _responderEncuesta(
-                    pregunta: _datosEncuesta.pregunta,
-                    respuestas: _datosEncuesta.respuestas,
+                    pregunta: _datosEncuesta!.pregunta,
+                    respuestas: _datosEncuesta!.respuestas,
                   ),
                   label: 'Responder',
                 ),
@@ -117,6 +117,17 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
           _noMolestar = estadoNoMolestar['valor'] == '1' ? 1 : 0;
         } else {
           _noMolestar = 3;
+        }
+      });
+    });
+
+    configUsuarioProvider
+        .obtenerEstadoConfig(_prefs.usuarioLogged, 7)
+        .then((estadoAccesos) {
+      if (!mounted) return;
+      setState(() {
+        if (estadoAccesos.containsKey('valor')) {
+          _qrResidente = estadoAccesos['valor'] == '1';
         }
       });
     });
@@ -187,7 +198,7 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
                   Text('Tema', style: TextStyle(fontSize: 10)),
                 ],
               ),
-              onPressed: MyApp.of(context).changeTheme),
+              onPressed: MyApp.of(context)!.changeTheme),
           const SizedBox(width: 10),
           _creaBtnContacto(),
           const SizedBox(width: 15),
@@ -264,10 +275,11 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
   }
 
   Widget _creaItemContacto(
-      {String title,
-      IconData icon,
+      {
+      required String title,
+      IconData? icon,
       String rutaIcon = '',
-      @required Function onPressed}) {
+      required Function() onPressed}) {
     return GestureDetector(
       child: Row(
         mainAxisSize: MainAxisSize.max,
@@ -284,7 +296,7 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
                 child: Text(title,
                     textAlign: TextAlign.center,
                     style: utils.estiloBotones(16,
-                        color: Theme.of(context).textTheme.bodyText2.color))),
+                        color: Theme.of(context).textTheme.bodyText2!.color!))),
           ),
           const SizedBox(width: 10),
           Container(
@@ -458,7 +470,7 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
     );
   }
 
-  Widget _creaQuintaFila() {
+   Widget _creaQuintaFila() {
     return Container(
       height: 120,
       padding: EdgeInsets.only(top: 20.0),
@@ -511,14 +523,19 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
   }
 
   Widget _creaBtnIcono(
-      {String rutaIcono, String titulo, String subtitulo, String ruta}) {
+      {required String rutaIcono,
+      required String titulo,
+      String? subtitulo,
+      required String ruta}) {
     return ElevatedContainer(
-      child: RaisedButton(
-          padding: EdgeInsets.symmetric(horizontal: 10),
-          elevation: 0,
-          color: Theme.of(context).cardColor,
-          shape:
-              RoundedRectangleBorder(borderRadius: BorderRadius.circular(15.0)),
+      child: ElevatedButton(
+          style: ElevatedButton.styleFrom(
+            primary: Theme.of(context).cardColor,
+            padding: const EdgeInsets.symmetric(horizontal: 10.0),
+            elevation: 0,
+            shape: RoundedRectangleBorder(
+                borderRadius: BorderRadius.circular(15.0)),
+          ),
           child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
             children: [
@@ -534,12 +551,17 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
                     titulo,
                     wrapWords: false,
                     style: utils.estiloBotones(30,
-                        color: Theme.of(context).textTheme.bodyText2.color),
+                        color: Theme.of(context).textTheme.bodyText2!.color!),
                   ),
                 ),
               ]),
               Visibility(
-                  visible: subtitulo != null, child: Text(subtitulo ?? ''))
+                  visible: subtitulo != null,
+                  child: Text(
+                    subtitulo ?? '',
+                    style: TextStyle(
+                        color: Theme.of(context).textTheme.bodyText2!.color!),
+                  ))
             ],
           ),
           onPressed: () => Navigator.pushNamed(context, ruta)),
@@ -547,18 +569,19 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
   }
 
   Widget _creaBtnIconoMini({
-    String rutaIcono,
-    String titulo,
-    String ruta,
-    Function onPressed,
+    required String rutaIcono,
+    required String titulo,
+    String? ruta,
+    Function()? onPressed,
   }) {
     return ElevatedContainer(
-      child: RaisedButton(
-          shape:
-              RoundedRectangleBorder(borderRadius: BorderRadius.circular(15.0)),
-          color: Theme.of(context).cardColor,
-          elevation: 0,
-          padding: EdgeInsets.zero,
+      child: ElevatedButton(
+          style: ElevatedButton.styleFrom(
+              shape: RoundedRectangleBorder(
+                  borderRadius: BorderRadius.circular(15.0)),
+              primary: Theme.of(context).cardColor,
+              elevation: 0,
+              padding: EdgeInsets.zero),
           child: Padding(
             padding: EdgeInsets.all(10.0),
             child: Column(
@@ -579,7 +602,7 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
                     maxLines: 2,
                     wrapWords: false,
                     style: utils.estiloBotones(16,
-                        color: Theme.of(context).textTheme.bodyText2.color),
+                        color: Theme.of(context).textTheme.bodyText2!.color!),
                   ),
                 ),
               ],
@@ -598,9 +621,10 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
       child: ElevatedContainer(
         child: ClipRRect(
           borderRadius: BorderRadius.circular(15.0),
-          child: FlatButton(
-              padding: EdgeInsets.only(left: 10.0, right: 10.0),
-              color: utils.colorAcentuado,
+          child: TextButton(
+              style: TextButton.styleFrom(
+                  padding: const EdgeInsets.only(left: 10.0, right: 10.0),
+                  backgroundColor: utils.colorAcentuado),
               child: Row(
                 mainAxisAlignment: MainAxisAlignment.spaceAround,
                 crossAxisAlignment: CrossAxisAlignment.center,
@@ -639,7 +663,8 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
     );
   }
 
-  _responderEncuesta({String pregunta, List<Respuesta> respuestas}) {
+  _responderEncuesta(
+      {required String pregunta, required List<Respuesta> respuestas}) {
     List<CupertinoActionSheetAction> actions = respuestas
         .map((element) => CupertinoActionSheetAction(
             child: Text(
@@ -661,7 +686,7 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
                 style: TextStyle(
                   fontSize: 24.0,
                   fontWeight: FontWeight.bold,
-                  color: Theme.of(context).textTheme.bodyText2.color,
+                  color: Theme.of(context).textTheme.bodyText2!.color,
                 ),
               ),
               actions: actions,
@@ -678,17 +703,20 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
             ));
   }
 
-  _enviarRespuesta(int respuesta) {
+  _enviarRespuesta(int? respuesta) {
     creaDialogProgress(context, 'Enviando respuesta...');
     avisosProvider
         .enviarRespuestaEncuesta(_prefs.usuarioLogged, respuesta)
         .then((resultados) {
       Navigator.of(context).pop('dialog');
       if (resultados.containsKey(1)) {
-        _scaffoldKey.currentState.showSnackBar(SnackBar(
+        ScaffoldMessenger.of(context).showSnackBar(SnackBar(
             content: Row(
               children: <Widget>[
-                Icon(Icons.send),
+                Icon(
+                  Icons.send,
+                  color:Theme.of(context).snackBarTheme.actionTextColor,
+                ),
                 SizedBox(
                   width: 15,
                 ),
@@ -702,11 +730,11 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
             action: SnackBarAction(
               label: 'Ver resultados',
               onPressed: () =>
-                  _mostrarResultados(_datosEncuesta.pregunta, resultados[1]),
+                  _mostrarResultados(_datosEncuesta!.pregunta, resultados[1]),
             )));
       } else {
-        Scaffold.of(context).showSnackBar(
-            utils.creaSnackBarIcon(Icon(Icons.error), resultados[2], 10));
+        ScaffoldMessenger.of(context).showSnackBar(
+            utils.creaSnackBarIcon(Icon(Icons.error, color: Theme.of(context).snackBarTheme.actionTextColor), resultados[2], 10));
       }
     });
   }
@@ -714,7 +742,8 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
   _mostrarResultados(String titulo, List<ResultadosEncuestaModel> resultados) {
     Map<String, double> dataMap = {};
     resultados.forEach((element) {
-      dataMap[element.respuestaEncuesta] = double.tryParse(element.porcentaje);
+      dataMap[element.respuestaEncuesta] =
+          double.tryParse(element.porcentaje) ?? 0.0;
     });
     //Map<String,double> dataMap={"Test":1,"Test2":1};
     creaDialogWidget(
@@ -816,8 +845,11 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
     Navigator.pop(context);
     setState(() {
       _noMolestar = valor;
-      Scaffold.of(context).showSnackBar(utils.creaSnackBarIcon(
-          Icon(resultado['OK'] == 1 ? Icons.notifications_active : Icons.error),
+      ScaffoldMessenger.of(context).showSnackBar(utils.creaSnackBarIcon(
+          Icon(
+            resultado['OK'] == 1 ? Icons.notifications_active : Icons.error,
+            color: Theme.of(context).snackBarTheme.actionTextColor
+          ),
           resultado['message'],
           5));
     });
@@ -825,10 +857,8 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
 
   _launchWhatsApp(String numero, String mensaje) async {
     final link = WhatsAppUnilink(phoneNumber: numero, text: mensaje);
-    // Convert the WhatsAppUnilink instance to a string.
-    // Use either Dart's string interpolation or the toString() method.
-    // The "launch" method is part of "url_launcher".
-    await launch('$link');
+    await launchUrl(Uri.parse(link.toString()),
+        mode: LaunchMode.externalApplication);
   }
 
   _abrirSoporte() {
@@ -852,7 +882,7 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
     if (codeResponse['statusCode'] == 201) {
       final fecha = DateTime.tryParse(codeResponse["vigencia"] ?? "");
       creaDialogQR(
-          _scaffoldKey.currentContext,
+          _scaffoldKey.currentContext!,
           '',
           Column(
             children: [
@@ -864,7 +894,7 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
           '',
           'Cancelar',
           () => {},
-          () => Navigator.of(_scaffoldKey.currentContext).pop('dialog'),
+          () => Navigator.of(_scaffoldKey.currentContext!).pop('dialog'),
           barrierDismissible: false,
           btnPos: false);
     } else {
