@@ -32,20 +32,20 @@ class _NuevoVisitanteRostroPageState extends State<NuevoVisitanteRostroPage> {
   final _txtApMatCtrl = TextEditingController();
   final visitanteProvider = VisitantesFreqProvider();
   final _configUsuarioProvider = ConfigUsuarioProvider();
-  String _seleccionTipoAcceso;
-  String _seleccionTipoVisitante;
+  String? _seleccionTipoAcceso;
+  String? _seleccionTipoVisitante;
   bool _registrando = false;
   bool _imagenLista = false, _mostrarErrorImg = false;
   bool _mostrarErrorAcceso = false;
   bool _visitanteRegistrado = false;
-  File _imgRostro;
+  File? _imgRostro;
   int _tipoRostro = 0;
   List arg = [];
   String _tipoAccesoVisitante = '';
   String _tipoAccesoColono = '';
-  String _tempTipoAccess;
-  List<TipoVisitanteModel> tipoVisita = new List();
-  List<DropdownMenuItem<String>> listTipo = new List();
+  late String _tempTipoAccess;
+  List<TipoVisitanteModel> tipoVisita = [];
+  List<DropdownMenuItem<String>> listTipo = [];
 
   @override
   void initState() {
@@ -70,7 +70,7 @@ class _NuevoVisitanteRostroPageState extends State<NuevoVisitanteRostroPage> {
 
   @override
   Widget build(BuildContext context) {
-    arg = ModalRoute.of(context).settings.arguments;
+    arg = ModalRoute.of(context)!.settings.arguments as List<dynamic>;
     _tipoRostro = arg[0];
     _tipoAccesoVisitante = arg[1]['acceso_visitante'];
     _tipoAccesoColono = arg[1]['acceso_colono'];
@@ -151,7 +151,7 @@ class _NuevoVisitanteRostroPageState extends State<NuevoVisitanteRostroPage> {
         labelText: label,
       ),
       validator: (texto) {
-        if (utils.textoVacio(texto))
+        if (utils.textoVacio(texto!))
           return 'Ingresa el nombre';
         else
           return null;
@@ -174,7 +174,7 @@ class _NuevoVisitanteRostroPageState extends State<NuevoVisitanteRostroPage> {
         labelText: label,
       ),
       validator: (texto) {
-        if (utils.textoVacio(texto))
+        if (utils.textoVacio(texto!))
           return 'Ingresa el apellido paterno';
         else
           return null;
@@ -196,7 +196,7 @@ class _NuevoVisitanteRostroPageState extends State<NuevoVisitanteRostroPage> {
         labelText: label,
       ),
       validator: (texto) {
-        if (utils.textoVacio(texto))
+        if (utils.textoVacio(texto!))
           return 'Ingresa el apellido materno';
         else
           return null;
@@ -213,9 +213,9 @@ class _NuevoVisitanteRostroPageState extends State<NuevoVisitanteRostroPage> {
           isExpanded: true,
           value: _seleccionTipoVisitante,
           items: listTipo,
-          onChanged: (opc) {
+          onChanged: (String? opc) {
             setState(() {
-              _seleccionTipoVisitante = opc;
+              _seleccionTipoVisitante = opc!;
             });
           },
         ),
@@ -248,9 +248,9 @@ class _NuevoVisitanteRostroPageState extends State<NuevoVisitanteRostroPage> {
                 : Theme.of(context).disabledColor,
           ),
           items: getOpcionesDropdown(),
-          onChanged: (opc) {
+          onChanged: (String? opc) {
             setState(() {
-              _seleccionTipoAcceso = opc;
+              _seleccionTipoAcceso = opc!;
             });
           },
         ),
@@ -316,7 +316,7 @@ class _NuevoVisitanteRostroPageState extends State<NuevoVisitanteRostroPage> {
       onTap: _registrando ? null : _mostrarOpcImagen,
       child: ElevatedContainer(
         child: ClipRRect(
-          borderRadius: BorderRadius.circular(20),
+          borderRadius: BorderRadius.circular(15),
           child: Container(
             width: 200,
             height: 250,
@@ -335,7 +335,7 @@ class _NuevoVisitanteRostroPageState extends State<NuevoVisitanteRostroPage> {
                   ? Container(
                       width: double.infinity,
                       child: Image.file(
-                        _imgRostro,
+                        _imgRostro!,
                         errorBuilder: (context, error, stackTrace) =>
                             Container(),
                         fit: BoxFit.cover,
@@ -364,7 +364,7 @@ class _NuevoVisitanteRostroPageState extends State<NuevoVisitanteRostroPage> {
       if (Platform.isAndroid) {
         if (!await utils.obtenerPermisosAndroid()) throw 'permission_denied';
       }
-      var imgFile = await picker.ImagePicker.pickImage(
+      var imgFile = await picker.ImagePicker().pickImage(
           source: source, maxHeight: 1024, maxWidth: 768, imageQuality: 50);
       if (imgFile != null) {
         var fixedImg = await utils.fixExifRotation(imgFile.path);
@@ -375,24 +375,40 @@ class _NuevoVisitanteRostroPageState extends State<NuevoVisitanteRostroPage> {
             _imagenLista = true;
           });
         } else {
-          _scaffoldKey.currentState.showSnackBar(utils.creaSnackBarIcon(
-              Icon(Icons.error), 'La imagen no está en formato vertical', 2));
+          ScaffoldMessenger.of(context).showSnackBar(utils.creaSnackBarIcon(
+              Icon(
+                Icons.error,
+                color: Theme.of(context).snackBarTheme.actionTextColor
+              ),
+              'La imagen no está en formato vertical',
+              2));
         }
       }
     } on PlatformException catch (e) {
       String mensajeError = utils.messageImagePlatformException(e);
-      _scaffoldKey.currentState.showSnackBar(utils.creaSnackBarIcon(
+      ScaffoldMessenger.of(context).showSnackBar(utils.creaSnackBarIcon(
           Icon(Icons.error),
           'Ocurrió un error al procesar la imagen. $mensajeError',
           2));
     } catch (e) {
-      String mensajeError = utils.messageErrorImage(e);
+      String mensajeError = utils.messageErrorImage(e as Exception);
       
-      _scaffoldKey.currentState.showSnackBar(utils.creaSnackBarIcon(
+      ScaffoldMessenger.of(context).showSnackBar(utils.creaSnackBarIcon(
           Icon(Icons.error),
           'Ocurrió un error al procesar la imagen. $mensajeError',
           2));
     }
+  }
+
+  Future<File> fixExifRotation(String imagePath) async {
+    final originalFile = File(imagePath);
+    final imgTools.Image capturedImage =
+        imgTools.decodeImage(await originalFile.readAsBytes())!;
+    final imgTools.Image orientedImage =
+        imgTools.bakeOrientation(capturedImage);
+    await originalFile
+        .writeAsBytes(imgTools.encodeJpg(orientedImage, quality: 50));
+    return originalFile;
   }
 
   Widget _creaTextoErrorImg() {
@@ -428,10 +444,13 @@ class _NuevoVisitanteRostroPageState extends State<NuevoVisitanteRostroPage> {
   }
 
   Widget _creaBtnRegistrar() {
-    return RaisedButton(
-        color: utils.colorAcentuado,
-        disabledColor: utils.colorSecundario,
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(15)),
+    return ElevatedButton(
+        style: ElevatedButton.styleFrom(
+          primary: utils.colorAcentuado,
+          onSurface: utils.colorSecundario,
+          shape:
+              RoundedRectangleBorder(borderRadius: BorderRadius.circular(15)),
+        ),
         child: Container(
             height: 60,
             width: double.infinity,
@@ -457,10 +476,10 @@ class _NuevoVisitanteRostroPageState extends State<NuevoVisitanteRostroPage> {
   }
 
   _submitForm() {
-    if (_formKey.currentState.validate() &&
+    if (_formKey.currentState!.validate() &&
         _seleccionTipoAcceso != null &&
         _imgRostro != null) {
-      _formKey.currentState.save();
+      _formKey.currentState!.save();
       _mostrarErrorImg = false;
       _mostrarErrorAcceso = false;
       _registrando = true;
@@ -487,9 +506,9 @@ class _NuevoVisitanteRostroPageState extends State<NuevoVisitanteRostroPage> {
         apPaterno: _txtApPatCtrl.text,
         apMaterno: _txtApMatCtrl.text,
         tipoAcceso: _seleccionTipoAcceso,
-        imgRostroB64: base64Encode(_imgRostro.readAsBytesSync()),
+        imgRostroB64: base64Encode(_imgRostro!.readAsBytesSync()),
         tipo: _tipoRostro,
-        tipoVisitante: _tipoRostro != 1 ? _seleccionTipoVisitante : "");
+        tipoVisitante: _tipoRostro != 1 ? _seleccionTipoVisitante! : "");
     switch (estatus['OK']) {
       case 1:
         //
