@@ -6,11 +6,11 @@ import 'package:dostop_v2/src/utils/preferencias_usuario.dart';
 import 'package:dostop_v2/src/utils/utils.dart' as utils;
 import 'package:flutter_svg/svg.dart';
 
-import 'package:flutter_swiper/flutter_swiper.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:flutter_swiper_plus/flutter_swiper_plus.dart';
 import 'package:pinch_zoom_image_last/pinch_zoom_image_last.dart';
 
 class VisitaDetallePage extends StatefulWidget {
@@ -21,9 +21,9 @@ class VisitaDetallePage extends StatefulWidget {
 class _VisitaDetallePageState extends State<VisitaDetallePage> {
   final ReportesProvider _reportesProvider = ReportesProvider();
   final _prefs = PreferenciasUsuario();
-  AppBar appBar;
-  double availableHeight;
-  List imagenes = [];
+  late AppBar appBar;
+  late double availableHeight;
+  List<Map<String, dynamic>> imagenes = [];
   bool loadImg = false;
 
   @override
@@ -35,7 +35,8 @@ class _VisitaDetallePageState extends State<VisitaDetallePage> {
   @override
   Widget build(BuildContext context) {
     appBar = utils.appBarLogo(titulo: 'Visita');
-    final VisitaModel _visita = ModalRoute.of(context).settings.arguments;
+    final VisitaModel _visita =
+        ModalRoute.of(context)!.settings.arguments as VisitaModel;
     availableHeight = MediaQuery.of(context).size.height -
         appBar.preferredSize.height -
         MediaQuery.of(context).padding.top -
@@ -74,9 +75,9 @@ class _VisitaDetallePageState extends State<VisitaDetallePage> {
                   builder:
                       (BuildContext context, AsyncSnapshot<List> snapshot) {
                     if (snapshot.hasData) {
-                      imagenes = snapshot.data;
+                      imagenes = snapshot.data! as List<Map<String, dynamic>>;
                       return _imagenesVisitante(
-                          visita.idVisitas, snapshot.data, visita);
+                          visita.idVisitas, imagenes, visita);
                     } else {
                       return Stack(children: [
                         Image.asset(utils.rutaGifLoadRed,
@@ -91,7 +92,8 @@ class _VisitaDetallePageState extends State<VisitaDetallePage> {
     );
   }
 
-  Widget _imagenesVisitante(String id, List<Map> imagenes, VisitaModel visita) {
+  Widget _imagenesVisitante(
+      String id, List<Map<String, dynamic>> imagenes, VisitaModel visita) {
     if (imagenes.length == 0)
       return Column(
         children: [
@@ -153,11 +155,9 @@ class _VisitaDetallePageState extends State<VisitaDetallePage> {
                                   _imgOrientacion(
                                       context, 240, imagenes[index]['img']),
                                   Text(
-                                      'Mantén presionada cualquier imagen para guardarla en tu galería.',
-                                      style: TextStyle(
-                                        fontSize: 12
-                                      ),
-                                      ),
+                                    'Mantén presionada cualquier imagen para guardarla en tu galería.',
+                                    style: TextStyle(fontSize: 12),
+                                  ),
                                   _datosVisitante(visita, context)
                                 ]),
                           );
@@ -174,16 +174,16 @@ class _VisitaDetallePageState extends State<VisitaDetallePage> {
       children: [
         const SizedBox(height: 30),
         Text(
-          visita.fechaEntrada.isNotEmpty && visita.fechaEntrada != null
+          visita.fechaEntrada.isNotEmpty
               ? '${utils.fechaCompleta(DateTime.tryParse(visita.fechaEntrada))} ${visita.horaEntrada}'
               : '',
           style: utils.estiloTextoSombreado(18, dobleSombra: false),
         ),
         const SizedBox(height: 10),
         Visibility(
-          visible: visita.fechaSalida.isNotEmpty && visita.fechaSalida != null,
+          visible: visita.fechaSalida != null && visita.fechaSalida!.isNotEmpty,
           child: Text(
-            'Salida: ${utils.fechaCompleta(DateTime.tryParse(visita.fechaSalida))} ${visita.horaSalida}',
+            'Salida: ${utils.fechaCompleta(DateTime.tryParse(visita.fechaSalida!))} ${visita.horaSalida}',
             style: utils.estiloTextoSombreado(14,
                 dobleSombra: false, fontWeight: FontWeight.w500),
           ),
@@ -301,9 +301,7 @@ class _VisitaDetallePageState extends State<VisitaDetallePage> {
           Text('Tipo', style: utils.estiloTituloInfoVisita(12)),
           Text(visita.tipoVisitante == '' ? 'Visita' : visita.tipoVisitante,
               style: utils.estiloTextoSombreado(16, dobleSombra: false)),
-          Visibility(
-            visible: visita.codigo == '',
-            child: SizedBox(height: 10)),
+          Visibility(visible: visita.codigo == '', child: SizedBox(height: 10)),
           Text(visita.codigo == '' ? 'Motivo' : '',
               style: utils.estiloTituloInfoVisita(12)),
           Text(visita.codigo == '' ? visita.motivoVisita : '',
@@ -321,10 +319,10 @@ class _VisitaDetallePageState extends State<VisitaDetallePage> {
         builder: (BuildContext context,
             AsyncSnapshot<Map<String, dynamic>> snapshot) {
           if (snapshot.hasData) {
-            if (snapshot.data.containsKey('OK')) {
-              switch (snapshot.data['OK']) {
+            if (snapshot.data!.containsKey('OK')) {
+              switch (snapshot.data!['OK']) {
                 case 1:
-                  ReporteModel reporte = snapshot.data['datos'];
+                  ReporteModel reporte = snapshot.data!['datos'];
                   return _creaFAB(reporte, null);
                 case 2:
                   return _creaFAB(null, datos);
@@ -344,9 +342,13 @@ class _VisitaDetallePageState extends State<VisitaDetallePage> {
         });
   }
 
-  Widget _creaFAB(ReporteModel reporte, List<String> datos) {
-    return RaisedButton(
-        color: utils.colorPrincipal,
+  Widget _creaFAB(ReporteModel? reporte, List<String>? datos) {
+    return ElevatedButton(
+        style: ElevatedButton.styleFrom(
+          primary: utils.colorPrincipal,
+          shape:
+              RoundedRectangleBorder(borderRadius: BorderRadius.circular(15)),
+        ),
         child: Container(
           height: 60,
           child: Row(
@@ -364,22 +366,19 @@ class _VisitaDetallePageState extends State<VisitaDetallePage> {
           ),
         ),
         onPressed: () => reporte == null
-            ? _abrirReportePage(context, datos)
+            ? _abrirReportePage(context, datos!)
             : Navigator.of(context)
-                .pushNamed('SeguimientoInc', arguments: reporte),
-        shape:
-            RoundedRectangleBorder(borderRadius: BorderRadius.circular(15.0)));
+                .pushNamed('SeguimientoInc', arguments: reporte));
   }
 
   _abrirReportePage(BuildContext context, List<String> datos) async {
-    final result =
-        await Navigator.of(context).pushNamed('Incidente', arguments: datos) ??
-            false;
-    if (result) setState(() {});
+    final result = await Navigator.of(context)
+        .pushNamed('Incidente', arguments: datos);
+    if (result as bool) setState(() {});
   }
 }
 
-Map<String, dynamic> getColorEstatus(String estatus) {
+Map<String, dynamic> getColorEstatus(String? estatus) {
   switch (estatus) {
     case 'Aceptada':
       return {'icono': Icons.check_circle, 'color': utils.colorAcentuado};
