@@ -1,6 +1,8 @@
 import 'dart:convert';
 
 import 'package:dostop_v2/src/utils/preferencias_usuario.dart';
+import 'package:package_info_plus/package_info_plus.dart';
+import '../utils/utils.dart';
 import 'constantes_provider.dart' as constantes;
 
 import 'package:http/http.dart' as http;
@@ -16,7 +18,7 @@ class LoginProvider {
       'contrasena': password,
     };
     try {
-      final resp = await http.post('${constantes.urlApp}/login.php', body: authData);
+      final resp = await http.post(Uri.parse('${constantes.urlApp}/login.php'), body: authData);
       List decodeResp = json.decode(resp.body);
       decodeResp[0].forEach((String k, dynamic v) => mapResp[k] = v);
     } catch (e) {
@@ -39,25 +41,31 @@ class LoginProvider {
 
 
   Future<Map<String, dynamic>> registrarTokenFCM(String dispositivo,
-      {String idUsuario, String token}) async {
+      {String? idUsuario, String? token}) async {
     if (_prefs.token == '')
       return {
         'OK': 0,
         'message':
-            'No se ha podido obtener el token desde FCM, revisar que los servicios de Google esten funcionando '
-                'y que los servicios internos de la aplicación se pueden conectar a FCM'
+            'No se ha podido obtener el token desde FCM, revisar que los servicios esten funcionando '
+                'y que los servicios internos de la aplicación se pueden conectar a Firebase'
       };
     Map<String, dynamic> mapResp = Map<String, dynamic>();
-    final authData = {
-      'id': idUsuario ?? _prefs.usuarioLogged,
-      'token': token ?? _prefs.token,
-      'dispositivo': dispositivo,
-      'fecha':
-          DateFormat('yyyy-MM-dd HH:mm:ss').format(DateTime.now()).toString()
-    };
     try {
+      final deviceData = await getDeviceData();
+      final infoApp = await PackageInfo.fromPlatform();
+      final authData = {
+        'id': idUsuario ?? _prefs.usuarioLogged,
+        'token': token ?? _prefs.token,
+        'dispositivo': dispositivo,
+        'fecha':
+            DateFormat('yyyy-MM-dd HH:mm:ss').format(DateTime.now()).toString(),
+        'versionApp': infoApp.version,
+        'SODevice': deviceData['os'],
+        'brand': deviceData['brand'],
+        'model': deviceData['nameModel'],
+      };
       final resp =
-          await http.post('${constantes.urlApp}/registrar_token_disp.php', body: authData);
+          await http.post(Uri.parse('${constantes.urlApp}/registrar_token_disp.php'), body: authData);
       List decodeResp = json.decode(resp.body);
       decodeResp[0].forEach((String k, dynamic v) => mapResp[k] = v);
     } catch (e) {
@@ -80,10 +88,10 @@ class LoginProvider {
 
   Future<bool> logout() async {
     try {
-      final resp = await http.post('${constantes.urlApp}/elimina_token.php',
+      final resp = await http.post(Uri.parse('${constantes.urlApp}/elimina_token.php'),
           body: {'token': _prefs.token});
-      Map decodeResp = json.decode(resp.body);
-      //print(decodeResp);
+      Map? decodeResp = json.decode(resp.body);
+      print(decodeResp);
       if (decodeResp == null) return false;
       if (decodeResp.containsKey('estatus')) {
         if (decodeResp['estatus'] !='3')
